tatyu

+++ ogameasure/device/Pfeiffer/tpg261_lan.py
@@ -15,8 +15,8 @@ class tpg261_lan():
         self.com.send(b"\x05")
         time.sleep(0.3)
         self.raw_p = self.com.recv().decode('utf-8').strip().split(',') 
-        pressure = numpy.float64(self.raw_p[-1])
-        return float(pressure)
+        pressure = float(self.raw_p[-1])
+        return pressure
 
     def tpg_status(self):
         self.com.send(b"PR1 \r\n")
